requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pillow==11.0.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
